+++ src/client/Main.client.lua
@@ -7,9 +7,11 @@ local Client = script.Parent
 local CameraController = require(Client:WaitForChild("CameraController"))
 local HudController = require(Client:WaitForChild("HudController"))
 local LevelUpController = require(Client:WaitForChild("LevelUpController"))
+local MetaProgressionController = require(Client:WaitForChild("MetaProgressionController"))
 
 CameraController.start()
 HudController.start()
 LevelUpController.start()
+MetaProgressionController.start()
 
 print(string.format("[goblin] Client booted for %s", GameConfig.gameName))